Study and Add: support typescript

+++ webpack/plugin/ArchivedPlugin.js
@@ -10,20 +10,22 @@ class ArchivedPlugin {
   }
   apply(compiler) {
     compiler.plugin('emit', function (compilation, callback) {
-      const archivedName = this.options.archivedPath || 'Archived';
       const outputPath = compiler.outputPath;
-      const archivedPath = path.join(outputPath, archivedName);
-      if (!fs.existsSync(archivedPath)) {
-        fs.mkdirSync(archivedPath)
-      }
-      const flagPath = path.join(archivedPath, moment().format('YYYYMMDDHHmmss'));
-      fs.mkdirSync(flagPath);
+      if (fs.existsSync(outputPath)) {
+        const archivedName = this.options.archivedPath || 'Archived';
+        const archivedPath = path.join(outputPath, archivedName);
+        if (!fs.existsSync(archivedPath)) {
+          fs.mkdirSync(archivedPath)
+        }
+        const flagPath = path.join(archivedPath, moment().format('YYYYMMDDHHmmss'));
+        fs.mkdirSync(flagPath);
 
-      const files = fs.readdirSync(outputPath);
-      const assetsFiles = files.filter(value => (value !== archivedName));
-      assetsFiles.forEach((filePath) => {
-        fs.renameSync(path.join(outputPath, filePath), path.join(flagPath, filePath));
-      });
+        const files = fs.readdirSync(outputPath);
+        const assetsFiles = files.filter(value => (value !== archivedName));
+        assetsFiles.forEach((filePath) => {
+          fs.renameSync(path.join(outputPath, filePath), path.join(flagPath, filePath));
+        });
+      }
       callback();
     })
   }
